More descriptive comment for the input poll.
The code self-documents the "how" of a certain process. Saying
that while(true) is to create an eternal loop is duplication.
The user can see that. But _why_ are we creating an eternal loop?
Polling input methods. But it is not too clear to the uninitiated
why we are doing that.